+++ CHANGES.md
@@ -1,3 +1,7 @@
+* Implemented --outSAMunmapped Within KeepPairs option to record unmapped mate adjacent to the mapped one, in case single-end alignments are allowed.
+  For multi-mappers, the unmapped mate will be recored mulitple times adjacent to the mappet mate of each alignment.
+
+
 STAR 2.5.0c 2015/12/23
 ======================